scripts/requirements.txt
# Opik Agent Optimizer requirements
opik>=1.0.0
opik-optimizer>=0.1.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
litellm>=1.0.0
